backend/requirements.txt
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.4.2
Flask==3.0.0
Flask-Cors==4.0.0
google-api-core==2.14.0
google-auth==2.23.4
google-cloud-core==2.3.3
google-cloud-datastore==2.18.0
google-cloud-firestore==2.13.1
googleapis-common-protos==1.61.0
grpcio==1.59.3
grpcio-status==1.59.3
gunicorn==21.2.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.2
proto-plus==1.22.3
protobuf==4.25.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pymongo==4.6.0
requests==2.31.0
rsa==4.9
urllib3==2.1.0
Werkzeug==3.0.1
